requirements.txt
transformers>=4.32.0
elasticsearch==7.17.9
python-docx==0.8.11
nltk==3.8.1
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=2.0.0
numpy==1.24.3
langchain==0.0.352
docx2txt==0.8
boto3>=1.26.0
opensearch-py>=2.0.0
requests-aws4auth==1.2.3
sentence-transformers==2.5.1
jinja2==3.1.3
python-dotenv>=0.19.0
pydantic-settings>=2.0.0
torch>=2.1.0

# New dependencies for MySQL and authentication
pymysql==1.1.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
bcrypt==4.0.1 